requirements.txt
backcall==0.2.0
decorator==5.0.6
greenlet==1.0.0
importlib-metadata==3.10.1
ipython==7.22.0
ipython-genutils==0.2.0
ipython-sql==0.4.0
jedi==0.18.0
numpy==1.20.2
pandas==1.2.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prettytable==0.7.2
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
Pygments==2.8.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
SQLAlchemy==1.4.7
sqlparse==0.4.1
traitlets==5.0.5
typing-extensions==3.7.4.3
wcwidth==0.2.5
zipp==3.4.1